# src/Panko.Client/openapi/panko-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Panko.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/cases": {
      "get": {
        "tags": [
          "Cases"
        ],
        "summary": "List recent Cases",
        "description": "Returns at most 500 Cases visible to the caller.",
        "operationId": "ListCases",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentCases"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cases"
        ],
        "summary": "Create a Case",
        "description": "Creates a durable Case. Supply Idempotency-Key on every attempt; the idempotencyKey body field is also accepted, and both values must match when supplied.",
        "operationId": "CreateCase",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/cases/{id}/inputs": {
      "get": {
        "tags": [
          "Cases"
        ],
        "summary": "List canonical Case inputs",
        "description": "Returns a bounded audit page of at most 500 inputs.",
        "operationId": "ListCaseInputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfCaseInput"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cases/{id}": {
      "get": {
        "tags": [
          "Cases"
        ],
        "operationId": "GetCaseFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseFile"
                }
              }
            }
          },
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePending"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "304": {
            "description": "Not Modified"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/pagerduty/incidents": {
      "get": {
        "tags": [
          "PagerDutyPullApi"
        ],
        "operationId": "GetRecentPagerDutyIncidents",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentPagerDutyIncidents"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/pagerduty/incidents/{id}/trigger": {
      "post": {
        "tags": [
          "PagerDutyPullApi"
        ],
        "operationId": "TriggerPagerDutyIncident",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseTriggerResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiSynthesis": {
        "required": [
          "status",
          "summary",
          "possibleContributors",
          "unknowns",
          "recommendedChecks",
          "crumbHash",
          "diagnoses",
          "summaryParts",
          "summaryReferences"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "possibleContributors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "unknowns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recommendedChecks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "crumbHash": {
            "type": [
              "null",
              "string"
            ]
          },
          "diagnoses": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/AiDiagnosis"
            }
          },
          "summaryParts": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/AiSummaryPart"
            }
          },
          "summaryReferences": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/AiSummaryReference"
            }
          }
        }
      },
...
      "CaseFile": {
        "required": [
          "caseId",
          "pagerDutyIncidentId",
          "serviceId",
          "recipeId",
          "recipeRevision",
          "title",
          "urgency",
          "pagerDutyState",
          "status",
          "openedAt",
          "updatedAt",
          "caseFileVersion",
          "deterministicSummary",
          "ai",
          "trail",
          "crumbs",
          "crumbSources",
          "links",
          "causalMarkers",
          "pattern"
        ],
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "pagerDutyIncidentId": {
            "type": [
              "null",
              "string"
            ]
          },
          "serviceId": {
            "type": "string"
          },
          "recipeId": {
            "type": "string"
          },
          "recipeRevision": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "urgency": {
            "type": "string"
          },
          "pagerDutyState": {
            "$ref": "#/components/schemas/PagerDutyIncidentState"
          },
          "status": {
            "type": "string"
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "caseFileVersion": {
            "type": "integer",
            "format": "int32"
          },
          "deterministicSummary": {
            "type": "string"
          },
          "ai": {
            "$ref": "#/components/schemas/AiSynthesis"
          },
          "trail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrailEntry"
            }
          },
          "crumbs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Crumb"
            }
          },
          "crumbSources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CrumbSourceStatus"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            }
          },
          "causalMarkers": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/CausalMarker"
            }
          },
          "pattern": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PatternContext"
              }
            ]
          },
          "origin": {
            "$ref": "#/components/schemas/CaseOrigin"
          },
          "inputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "projectedInputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CaseInput": {
        "required": [
          "id",
          "sequence",
          "inputVersion",
          "clientCrumbId",
          "producerPrincipal",
          "receivedAt",
          "occurredAt",
          "kind",
          "category",
          "severity",
          "summary",
          "excerpt",
          "declaredSource",
          "sourceReference",
          "url",
          "actor",
          "objectType",
          "objectId",
          "attributes",
          "trustLevel",
          "supersedesCrumbId",
          "supersededByCrumbId",
          "retractedAt",
          "retractedInputVersion",
          "active",
          "projectedInInputVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "inputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "clientCrumbId": {
            "type": "string"
          },
          "producerPrincipal": {
            "type": "string"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "kind": {
            "$ref": "#/components/schemas/SubmittedCrumbKind"
          },
          "category": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "excerpt": {
            "type": [
              "null",
              "string"
            ]
          },
          "declaredSource": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceReference": {
            "type": [
              "null",
              "string"
            ]
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          },
          "actor": {
            "type": [
              "null",
              "string"
            ]
          },
          "objectType": {
            "type": [
              "null",
              "string"
            ]
          },
          "objectId": {
            "type": [
              "null",
              "string"
            ]
          },
          "attributes": {
            "type": "object"
          },
          "trustLevel": {
            "type": "string"
          },
          "supersedesCrumbId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "supersededByCrumbId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "retractedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "retractedInputVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "active": {
            "type": "boolean"
          },
          "projectedInInputVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          }
        }
      },
...
      "CaseOriginKind": {
        "enum": [
          "pagerDuty",
          "agent",
          "manual"
        ]
      },
      "CasePending": {
        "required": [
          "caseId",
          "pagerDutyIncidentId",
          "serviceId",
          "recipeId",
          "title",
          "urgency",
          "pagerDutyState",
          "status",
          "openedAt",
          "updatedAt",
          "caseFileVersion"
        ],
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "pagerDutyIncidentId": {
            "type": [
              "null",
              "string"
            ]
          },
          "serviceId": {
            "type": "string"
          },
          "recipeId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "urgency": {
            "type": "string"
          },
          "pagerDutyState": {
            "$ref": "#/components/schemas/PagerDutyIncidentState"
          },
          "status": {
            "type": "string"
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "caseFileVersion": {
            "type": "integer",
            "format": "int32"
          },
          "origin": {
            "$ref": "#/components/schemas/CaseOrigin"
          },
          "inputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "projectedInputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CaseTriggerResult": {
        "required": [
          "caseId",
          "caseUrl",
          "duplicate"
        ],
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "caseUrl": {
            "type": "string"
          },
          "duplicate": {
            "type": "boolean"
          }
        }
      },
...
      "CreateCaseRequest": {
        "required": [
          "recipeId",
          "title",
          "serviceId",
          "urgency",
          "referenceTime"
        ],
        "type": "object",
        "properties": {
          "recipeId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "urgency": {
            "type": "string"
          },
          "referenceTime": {
            "type": "string",
            "format": "date-time"
          },
          "labels": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "idempotencyKey": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateCaseResponse": {
        "required": [
          "caseId",
          "origin",
          "inputVersion",
          "projectedInputVersion",
          "caseFileVersion",
          "status",
          "caseUrl"
        ],
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "origin": {
            "$ref": "#/components/schemas/CaseOriginKind"
          },
          "inputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "projectedInputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "caseFileVersion": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "caseUrl": {
            "type": "string"
          },
          "duplicate": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "PageOfCaseInput": {
        "required": [
          "total",
          "items"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseInput"
            }
          }
        }
      },
...
      "PagerDutyIncidentState": {
        "enum": [
          "triggered",
          "acknowledged",
          "escalated",
          "reassigned",
          "resolved",
          "unknown"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RecentCase": {
        "required": [
          "caseId",
          "origin",
          "recipeId",
          "serviceId",
          "title",
          "status",
          "inputVersion",
          "projectedInputVersion",
          "caseFileVersion",
          "createdBy",
          "updatedAt",
          "caseUrl"
        ],
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "origin": {
            "$ref": "#/components/schemas/CaseOriginKind"
          },
          "recipeId": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "inputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "projectedInputVersion": {
            "type": "integer",
            "format": "int64"
          },
          "caseFileVersion": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "caseUrl": {
            "type": "string"
          }
        }
      },
      "RecentCases": {
        "required": [
          "total",
          "cases"
        ],
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "cases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentCase"
            }
          }
        }
      },
...
      "RecentPagerDutyIncidents": {
        "required": [
          "since",
          "until",
          "hasMore",
          "incidents"
        ],
        "type": "object",
        "properties": {
          "since": {
            "type": "string",
            "format": "date-time"
          },
          "until": {
            "type": "string",
            "format": "date-time"
          },
          "hasMore": {
            "type": "boolean"
          },
          "incidents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecentPagerDutyIncident"
            }
          }
        }
      },
...
      "SourceLink": {
        "required": [
          "label",
          "url"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "SubmittedCrumbKind": {
        "enum": [
          "event",
          "crumb",
          "note"
        ]
      },
...
      "TrailEntry": {
        "required": [
          "id",
          "occurredAt",
          "source",
          "kind",
          "summary",
          "severity",
          "url",
          "actor",
          "objectType",
          "objectId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          },
          "actor": {
            "type": [
              "null",
              "string"
            ]
          },
          "objectType": {
            "type": [
              "null",
              "string"
            ]
          },
          "objectId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
...
}